Brendon Disrud (spawning update)
this allows for multiple spawnpoints

updated files are:
*     ListAdder
*     SpawnEnemy
*     EnemyPosition

How to use
put list adder on any object and specify a number for it. that object
will be removed form the world on game play.  Put a SpawnEnemy script on
an object and specify the number.  when the player walks through a
spawnenemy, it will trigger enemies to spawn at all of the points that a
"listadder" was for the same specified index (supported numbers 1
through 20)